+++ backend/utils/math.go
@@ -2,7 +2,7 @@ package utils
 
 import "math"
 
-func GetDiff(fromX, fromZ, toX, toZ float32) (float64, float64) {
+func GetDiff(fromX, fromZ, toX, toZ float64) (float64, float64) {
 	return float64(toX - fromX), float64(toZ - fromZ)
 }
